+++ src/math/Matrix.java
@@ -1,14 +1,14 @@
 package math;
 
 public class Matrix {
-	double matrixValues[][] = null;
+	float matrixValues[][] = null;
 
 	public Matrix(int m, int n) {
-		matrixValues = new double[m][n];
+		matrixValues = new float[m][n];
 	}
 
-	public Matrix(double values[][]) {
-		matrixValues = new double[values.length][values[0].length];
+	public Matrix(float values[][]) {
+		matrixValues = new float[values.length][values[0].length];
 		for (int i=0; i<values.length; i++) {
 			for (int j=0; j<values[0].length; j++) {
 				matrixValues[i][j] = values[i][j];
@@ -24,11 +24,11 @@ public class Matrix {
 		return matrixValues[0].length;
 	}
 
-	public double getValue(int i, int j) {
+	public float getValue(int i, int j) {
 		return matrixValues[i][j];
 	}
 
-	public void setValue(int i, int j, double value) {
+	public void setValue(int i, int j, float value) {
 		matrixValues[i][j] = value;
 	}
